requirements.txt
# Core Dependencies
google-adk[full]>=0.1.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0,<3.0.0
typing-extensions>=4.0.0
deprecated>=1.2.0

# Required for timezone support
backports.zoneinfo;python_version<"3.9"

# Development
tox>=3.28.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0